int \neq float

+++ Traction Control 2025:26/Classic Control/Slip.h
@@ -1,4 +1,4 @@
 #ifndef SLIP_H
 #define SLIP_H
-int Slip(int RLSpeed, int RRSpeed, int vehicleSpeed);
+float Slip(float RLSpeed, float RRSpeed, float vehicleSpeed);
 #endif
